+++ jpa_demo/src/main/java/com/springboot/jpa_demo/datasource1/repository/GymRepository2.java
@@ -3,6 +3,6 @@ package com.springboot.jpa_demo.datasource1.repository;
 import com.springboot.jpa_demo.datasource1.domain.Gym;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GymRepository2 extends CrudRepository<Gym,Long> {
+public interface GymRepository2 extends CrudRepository<Gym, Integer> {
    //有count方法,exist方法
 }
