+++ jpa_demo/src/main/java/com/springboot/jpa_demo/datasource1/repository/UserRepositpry.java
@@ -4,5 +4,5 @@ import com.springboot.jpa_demo.datasource1.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepositpry extends JpaRepository<User,Integer> {
-    public User findById(Long id);
+    public User findById(int id);
 }
